lots of refactoring

+++ app/Http/Controllers/API/V1/BeaconProvidersController.php
@@ -9,39 +9,34 @@ use App\Http\Resources\BeaconProviderResource;
 use App\Models\BeaconProvider;
 use Exception;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
 use Illuminate\Validation\ValidationException;
 
 class BeaconProvidersController extends Controller
 {
-    /**
-     * BeaconsController constructor.
-     */
-    public function __construct()
-    {
-        $this->middleware('permission:beacon-providers.create')->only(['store']);
-        $this->middleware('permission:beacon-providers.read')->only(['index', 'paginated', 'show']);
-        $this->middleware('permission:beacon-providers.update')->only(['update']);
-        $this->middleware('permission:beacon-providers.delete')->only(['destroy', 'bulkDestroy']);
-    }
 
     /**
      * @return JsonResponse
+     * @throws Exception
      */
     public function index()
     {
+        $this->authorize('viewAny', BeaconProvider::class);
+
         $providers = BeaconProvider::all();
 
-        return $this->json($providers, Response::HTTP_OK);
+        return $this->json(BeaconProviderResource::collection($providers), Response::HTTP_OK);
     }
 
     /**
      * @return AnonymousResourceCollection
+     * @throws Exception
      */
     public function paginated()
     {
+        $this->authorize('viewAny', BeaconProvider::class);
+
         $providers = BeaconProvider::query()
             ->jsonPaginate($this->paginationNumber());
 
@@ -70,37 +65,34 @@ class BeaconProvidersController extends Controller
             ]);
         }
 
-        return $this->json($provider, Response::HTTP_CREATED);
+        return $this->json(new BeaconProviderResource($provider), Response::HTTP_CREATED);
     }
 
     /**
      * @param BeaconProvider $provider
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function show(BeaconProvider $provider)
     {
-        return $this->json($provider, Response::HTTP_OK);
+        $this->authorize('view', BeaconProvider::class);
+
+        return $this->json(new BeaconProviderResource($provider), Response::HTTP_OK);
     }
 
     /**
-     * @param Request $request
+     * @param BeaconProviderRequest $request
      * @param BeaconProvider $provider
      *
      * @return JsonResponse
-     * @throws ValidationException
+     * @throws Exception
      */
-    public function update(Request $request, BeaconProvider $provider)
+    public function update(BeaconProviderRequest $request, BeaconProvider $provider)
     {
-        $this->validate($request, [
-            'name' => 'required|max:255'
-        ]);
+        $provider->fill($request->validated())->save();
 
-        $provider->fill([
-            'name' => $request->get('name')
-        ])->save();
-
-        return $this->json($provider, Response::HTTP_OK);
+        return $this->json(new BeaconProviderResource($provider), Response::HTTP_OK);
     }
 
     /**
@@ -111,6 +103,8 @@ class BeaconProvidersController extends Controller
      */
     public function destroy(BeaconProvider $provider)
     {
+        $this->authorize('delete', BeaconProvider::class);
+
         $provider->delete();
 
         return $this->json(null, Response::HTTP_OK);
@@ -120,9 +114,12 @@ class BeaconProvidersController extends Controller
      * @param BulkDeleteRequest $request
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function bulkDestroy(BulkDeleteRequest $request)
     {
+        $this->authorize('delete', BeaconProvider::class);
+
         collect($request->get('items'))->each(function ($provider) {
             if ($providerToDelete = BeaconProvider::find($provider['id'])) {
                 $providerToDelete->delete();
