+++ app/Http/Controllers/API/V1/RulesController.php
@@ -5,8 +5,9 @@ namespace App\Http\Controllers\API\V1;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\RuleRequest;
 use App\Http\Resources\RuleResource;
-use App\Models\Beacon;
 use App\Models\Container;
+use App\Models\Rule;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 
@@ -14,22 +15,12 @@ class RulesController extends Controller
 {
 
     /**
-     * BeaconsController constructor.
-     */
-    public function __construct()
-    {
-        $this->middleware(['permission:beacons.create', 'permission:containers.create'])->only(['store']);
-        $this->middleware(['permission:beacons.read', 'permission:containers.read'])->only(['show']);
-        $this->middleware(['permission:beacons.update', 'permission:containers.update'])->only(['update']);
-        $this->middleware(['permission:beacons.delete', 'permission:containers.delete'])->only(['destroy']);
-    }
-
-    /**
      * @param RuleRequest $request
      * @param Container $container
      * @param $beaconId
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function store(RuleRequest $request, Container $container, $beaconId)
     {
@@ -46,9 +37,12 @@ class RulesController extends Controller
      * @param $rule
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function show(Container $container, $beaconId, $rule)
     {
+        $this->authorize('view', Rule::class);
+
         $beacon = $container->beacons()->findOrFail($beaconId);
 
         $rule = $beacon->pivot->rules()->findOrFail($rule);
@@ -63,6 +57,7 @@ class RulesController extends Controller
      * @param $rule
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function update(RuleRequest $request, Container $container, $beaconId, $rule)
     {
@@ -81,9 +76,12 @@ class RulesController extends Controller
      * @param $rule
      *
      * @return JsonResponse
+     * @throws Exception
      */
     public function destroy(Container $container, $beaconId, $rule)
     {
+        $this->authorize('delete', Rule::class);
+
         $beacon = $container->beacons()->findOrFail($beaconId);
 
         $beacon->pivot->rules()->findOrFail($rule)->delete();
