+++ app/Http/Controllers/API/V1/TagsController.php
@@ -8,7 +8,10 @@ use App\Http\Requests\TagRequest;
 use App\Http\Resources\TagResource;
 use App\Models\Tag;
 use App\Services\Models\TagService;
+use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
 
 class TagsController extends Controller
@@ -26,22 +29,19 @@ class TagsController extends Controller
      */
     public function __construct(TagService $tagService)
     {
-        $this->middleware('permission:tags.create')->only(['store']);
-        $this->middleware('permission:tags.read')->only(['index', 'show', 'paginated']);
-        $this->middleware('permission:tags.update')->only(['update']);
-        $this->middleware('permission:tags.delete')->only(['destroy', 'bulkDestroy']);
-
         $this->tagService = $tagService;
     }
-
 
     /**
      * @param Request $request
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
+     * @throws Exception
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Tag::class);
+
         $tags = Tag::query()
             ->filter($request)
             ->get();
@@ -52,10 +52,13 @@ class TagsController extends Controller
     /**
      * @param Request $request
      *
-     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection
+     * @return AnonymousResourceCollection
+     * @throws Exception
      */
     public function paginated(Request $request)
     {
+        $this->authorize('viewAny', Tag::class);
+
         $tags = Tag::query()
             ->filter($request)
             ->jsonPaginate($this->paginationNumber());
@@ -66,11 +69,12 @@ class TagsController extends Controller
     /**
      * @param TagRequest $request
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
+     * @throws Exception
      */
     public function store(TagRequest $request)
     {
-        $tag = $this->tagService->create($request);
+        $tag = $this->tagService->create($request->validated());
 
         return $this->json(new TagResource($tag), Response::HTTP_CREATED);
     }
@@ -78,10 +82,13 @@ class TagsController extends Controller
     /**
      * @param Tag $tag
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
+     * @throws Exception
      */
     public function show(Tag $tag)
     {
+        $this->authorize('view', Tag::class);
+
         return $this->json(new TagResource($tag), Response::HTTP_OK);
     }
 
@@ -89,11 +96,12 @@ class TagsController extends Controller
      * @param TagRequest $request
      * @param Tag $tag
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
+     * @throws Exception
      */
     public function update(TagRequest $request, Tag $tag)
     {
-        $tag = $this->tagService->update($tag, $request);
+        $tag = $this->tagService->update($tag, $request->validated());
 
         return $this->json(new TagResource($tag), Response::HTTP_OK);
     }
@@ -101,11 +109,13 @@ class TagsController extends Controller
     /**
      * @param Tag $tag
      *
-     * @return \Illuminate\Http\JsonResponse
-     * @throws \Exception
+     * @return JsonResponse
+     * @throws Exception
      */
     public function destroy(Tag $tag)
     {
+        $this->authorize('delete', Tag::class);
+
         $tag->delete();
 
         return $this->json(null, Response::HTTP_OK);
@@ -114,10 +124,13 @@ class TagsController extends Controller
     /**
      * @param BulkDeleteRequest $request
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
+     * @throws Exception
      */
     public function bulkDestroy(BulkDeleteRequest $request)
     {
+        $this->authorize('delete', Tag::class);
+
         collect($request->get('items'))->each(function ($tag) {
             if ($tagToDelete = Tag::find($tag['id'])) {
                 $tagToDelete->delete();
